requirements.txt
paho-mqtt>=1.4.0
PyYAML